Cleanup: remove #ifver blocks for old Wesnoth versions (<= 1.13)
We no longer support Wesnoth 1.13 and earlier (new code is written
with the assumption that Wesnoth is at least 1.14),
so those backward compatibility blocks are no longer needed.

Note: utils/stats.cfg was skipped, because there is a plan to replace
this file with Lua anyway.

Note: indent of "#ifndef EASY" was manually fixed in 10_The_Poison.cfg,
because it confused #ifver replacement script.