inventory: implement "Unit information" and "Pick up items on the ground"
These screens are shown in custom events, we can invoke them from Lua.

The callbacks in inventory_config.action_buttons can now be:
1) onclick - doesn't close the dialog.

Perfect for saying funny things in Tutorial scenario,
but can also be used together with multipage widget
to switch to screens like Item Storage or Manage Retaliation Attacks
without actually closing the Inventory dialog.

2) onsubmit - closes the dialog.
When we no longer need the dialog, or when the UI we plan to display
is better without the modal dialog hanging below it.
For example, trying to use onclick handler for "Show unit information"
results in this button being shown on top of the information screen.

3) default callback is only called for buttons that are neither onclick
nor onsubmit, and it is a non-submitting callback by itself.

See https://github.com/Dugy/Legend_of_the_Invincibles/issues/145