inventory: refactoring: split unit-dependent and unit-independent logic
We now have the following:

1) methods like get_items_tab() and get_retaliation_tab() return WML of
the tab widget. The result is immutable and doesn't depend on the unit.

2) methods like onshow_items_tab() populate the tab with unit-specific
information. They can be called again if the data needs to change,
for example, when player changed the unit's sword and returned back to
the inventory screen (assuming Item Storage will be another tab).

It should further be refactored to support showing another unit without
actually closing the dialog. (for "items on units on recall list").

The button "Select weapons for retaliation" now correctly opens a tab
for this feature.