inventory: retaliation tab: hide "attack only" weapons like whirlwind
Because some attacks are "attack only" by design (e.g. whirlwind),
we shouldn't show them in this dialog, as they shouldn't be enabled.

We now keep a list of attacks that were disabled by this dialog.
If some attack is disabled but not present in the list, it's not shown.

This is backward compatible with WML version (same list format).

See https://github.com/Dugy/Legend_of_the_Invincibles/issues/145