inventory: redraw the Items screen after "Unequip all items"
Also set_dialog_callback() was moved away from onshow_items_tab(),
because Wesnoth hangs on some kind of race condition
when you try to replace the callback that is currently running.

In this case, callback of "Unequip all items" called goto_tab() on
the items tab, which called onshow_items_tab(),
which (previously) was modifying the callback of "Unequip all items".