inventory: disable non-applicable buttons for units on recall list
Because the unit on recall list isn't really on the map, it can't be
a receiver of "item pick" event (not to mention it would be illogical
for unit who isn't here to take something from the Item Storage).

So we disable things like Item Storage or Crafting for absent units.