inventory: handle the leftover items
Currently equipped leftover item is shown when you click on the slot.

This shows item storage of the correct sort (i.e. other swords if the
leftover item is a sword), but "Equip" button is hidden,
and the explanation "Why you can't equip these" is shown.

Leftover items can be unequipped, there is a warning "Are you sure".
In this warning, leftover items are called "purely nostalgic".

 See https://github.com/Dugy/Legend_of_the_Invincibles/issues/145 for details.