items.lua: exclude recently picked gems from list_regular()
The only reason gems are stored as items is because [object] tag
is used to show "which gem was picked" when they are picked.

Normally these gems are removed from the unit by update_stats(),
but it's not called when picking a gem (because there is no reason).

So we filter them here, so that they wouldn't be displayed
in "leftover" slot of the inventory dialog.