inventory: use a smaller tooltip (near the mouse, not below the dialog)
Tooltip below the dialog (tooltip_large) is sometimes not fully erased
when leaving the "items" tab for "storage" tab, and then returning
to the "items" tab.

(part of the old tooltip will be seen below the dialog - until player
hovers the mouse over some other slot)

But smaller tooltip (that is displayed on top of the dialog)
seems to be properly erased.

See #145 for details.