items.lua: don't add gold to Item Storage, just give it to current side
This solves #188.

Alternative was to not pick gold automatically when scenario is won.
Could be disappointing when you didn't pick the gold due to assuming
that it would be picked.

Note: because this happens at the end of scenario, player will only get
a carryover percentage of this gold in the next scenario.