Make Crafting dialog more compact
This is a partial fix for #206.
Width of column 1 and horizontal margins are slightly decreased
to make space for column 2 (list of recipes) and column 3 (information
about item and the number of gems).

This still doesn't fix the situation when "information about item"
is truncated. That label has a tonn of space, yet doesn't use it.
To be investigated.