Unbreak the Crafting window, step 2
1) Item type is now a horizontal listbox.
Surprisingly it looks ok in terms of UX.

2) Three columns are now as follows:
column1: listbox with recipes,
column2: gem statistics,
column3: item description

3) The trick is: now only the last column (column 3, item description)
changes its width when we select a new recipe,
which means that (in most cases) width/height of listboxes doesn't need
to be recalculated (which Wesnoth sometimes does imperfectly).

4) Although the width of first column changes when switching between
weapons and armours, it no longer affects other listboxes (gem stats
are almost always the same width, and item description has a tonn of
space, being in the horizontal_grow=true column and all),
so it now looks OK when we switch between Armours and Weapons.

(FIXME: [horizontal_listbox] with types causes Wesnoth to print warning
"error gui/layout: Failed to fit horizontal list to requested rect",
but it seems to be working correctly despite of this.
We are not adding a tonn of scrollbar-handling code to a widget that
has an automatically added scrollbar already
- not to "just suppress a warning", that is).

See issue #206 and issue #209 for why this change was needed.