inventory: recall: pressing Enter should work as if "View" was clicked
This can later be used for "Equip" button on Storage tab, etc.

Implementation details:
1) open_tab() now reopens the dialog if it has already been closed.
(as known from the boolean flag inventory_dialog.is_opened)

2) catch_enter_or_ok(field_id, callback) sets the callback that will be
called if the dialog was closed with Enter key or OK button.

Because get_dialog_value() can't be called after the dialog has
already been closed, catch_enter_or_ok() takes the ID of dialog widget
as the first parameter, and remembers its value in postshow()
and then passes this value to callback.

3) Recall tab uses catch_enter_or_ok() for its View handler.
To avoid code duplication, we just give id="ok" to the View button,
so this callback is called for View button too.

4) Because the dialog can now be reopened, inventory_dialog.mpsafety is
no longer constructed on every opening of the dialog.
Instead, mpsafety.export() clears the operations log, and is called by
top-level open_inventory_dialog(), which is not used for reopening.

5) Method open_inventory_dialog_unsynced() has now been renamed to
inventory_dialog.reopen_unsynced().
It can start with a clear mpsafety log or continue an existing log,
and doesn't even need to know it.

See #145 for details.