unitdata: greatly simplify loti.unit.effects()
There is no need to list advancements, items and other modifications
separately, because they are all handled in a very similar manner.

We also don't need to recalculate defence of crafted non-chest armours,
since it has nothing to do with effects (defence is property of item).

We also don't seem to need get_type_advancement(), because it's exactly
the same as the [advancement] tag within [modifications],
and we already have this tag.

The iterator code is now basically two loops:
1) loop over modification tags,
2) loop over [effect] tags within the current modification.

See issue #204.