Travis: workaround for latest tag being unknown due to git clone depth
See https://github.com/travis-ci/travis-ci/issues/7422 for details.

It wouldn't be wise to clone without --depth for a large repository,
but it will be a very small difference in this case (a couple of extra
seconds, maybe).
Actual tests (especially wmllint) take ~7 minutes.
So it should be ok for now.