Support non-calculated description (gems, etc.) in describe_item()
This fixes a regression bug of the change
"Use loti.unit.item_with_set_effects() in loti.item.describe_item()".

Some items (like gems) have a fixed, non-calculated description.
If it is present in item_list.cfg, then it should always be used.

The reason why this code was previously removed is because it was
also used for caching calculated descriptions,
and such cache became obsolete due to use of item_with_set_effects().