inventory: Tutorial: open new inventory dialog on step 4 (incomplete)
This demonstrates how the Inventory dialog can be used in the Tutorial
while still having control of what pressing certain buttons does
(in this case - "Crafting", "Item storage" and "Unit information"
are not allowed on this step of the Tutorial).

This is incomplete, as the unequipping itself is not handled yet.
Furthermore, Efraim has to pick another sword next.

We should also avoid confusion of the buttons "Unequip all items"
and "Pick up items on the ground" being active
(especially since "Unequip all items" does exactly the same as
what we asked player to do: Efraim has only the sword),
either by hiding them or by adding dialogue to them too.

See #145 for details.