Reorder mpsafety:queue() for easier debugging
This is for detecting possible errors in multiplayer_safety.lua.

Previously execute_immediately() was called on non-encoded operation,
meaning that in singleplayer we would never notice if the encoding was
performed incorrectly. (and would get unexpected OoS in multiplayer,
which would be difficult to troubleshoot)

Now we call execute_immediately() on an already encoded operation,
ensuring that local player and remote players work on the same data.