unitdata: implement Lua storage (incomplete)
See issue #204 for details.
This is a basis of Lua arrays-based implementation of unitdata methods.

What is not yet implemented:
1) effect_containers(unit) - not documented in API + not in testsuite,
2) on_equip() is not called yet (don't want to copy/paste the code from
WML implementation, it should be made into a separate function),
3) Import feature: when save game is loaded, all WML [modification]
tags should be imported into the new Lua array and removed from units.
4) Save/Load: implement wesnoth.persistent_tags.<something>.write
to serialize "global_storage" array from unitdata1.lua
(which would record it in the save file),
and then a similar reading function.
5) see FIXME and TODO comments.

NOTE: values returned by effects() are ordered differently from how
the WML implementation orders them (see effects() in unitdata1.lua).

If ordering is adjusted in testfile (unitdata.test.lua),
then loti.testsuite() already reports success on all tests.

TODO: modify the test, so that the order of effects() would be definitive,
not based on the order of items being equipped, etc.