Fix items not being picked up when winning the scenario
This is a regression bug.

Before item manipulations were migrated to Lua, "on the ground" items
sometimes contained "turn" parameter, which was used in "victory" event
to automatically pick up items that were dropped on the last turn.

Solution is to remember "turn" in loti.item.on_the_ground.add().