(inventory) Support "Drop" and "Destroy" for items currently on unit
This fixes a "lost feature that existed before" regression.

Previous (non-GUI) inventory dialog allowed to drop currently equipped
item to the ground. But new dropdown menu with Drop/Destroy only allows
to drop/destroy the item currently in storage.

Because storage is not accessible after turn 2, it became impossible to
unequip the item in a way that would leave it on the ground.

Solution is to add dropdown menu to Unequip button too, where the
buttons "Drop" and "Destroy" would be applied to equipped item.