Fix Item Storage being accessible after turn 2 after doing the Tutorial
A player has noticed that Item Storage is accessible after turn 2.
This wasn't reproduced in Gladiators, etc.

Turns out, we were relying on wesnoth.get_variable("tutorial.progress")
to determine "are we in Tutorial?" (during the Tutorial, such access to
the Item Storage should be allowed),
but this variable wasn't cleared at the end of Tutorial scenario,
so it existed in all the following scenarios. (lifting the restriction)