RARE_ITEM -> INVENTORY_ITEM
I don't have a save to test this against, but I think it's just a simple oversight.

I went to Nightmare Cellar and left on turn 1.  All of the loot was automatically picked up for me.  Since every other scenario in ch9 uses INF_ITEM, I bet this is the fix.